changed the interval for the enemy spawner to a float so you can make it less than one, to create enemeis faster than 1 per second

+++ Assets/Scripts/EnemySpawner.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class EnemySpawner : MonoBehaviour {
 	public bool spawning;
-	public int spawnInterval;
+	public float spawnInterval;
 
 	private bool started = false;
